pyproject.toml
[project]
name = "fashion-frenzy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.8.0",
    "google>=3.0.0",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "nest-asyncio>=1.6.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "streamlit>=1.54.0",
]
